refactor: move job & queue to resource dir
